+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased - Will be 2.0.0]
 ### Added
 - New WP-CLI command to manage loggers (list, start, pause, clean, purge and remove): see `wp help decalog logger` for details.
+- New WP-CLI command to manage listeners (list, enable, disable and auto-listening on/off): see `wp help decalog listener` for details.
 ### Changed
 - The PHP listener now takes care of activated/deactivated modules between web server and command-line configurations.
 ### Fixed